requirements.txt
fastapi
uvicorn
pyjwt
passlib[bcrypt]
sqlalchemy
psycopg2-binary
alembic
python-decouple
python-slugify
fastapi-mail
databases